+++ NEWS.md
@@ -2,7 +2,8 @@
 
 * Add support for datetimes
 * Add 'hms' to Suggests
-* Add new example that lets user uplaod data
+* Add new example that lets user upload data
+* Fix bug when categorical column is completely missing
 
 # shinydataviewer 0.1.0